feat: 补充V146下GroupInfo错乱的升级修复代码 (#1297)
:( 先行使特权了

+++ model/group_info.go
@@ -8,7 +8,7 @@ import (
 // GroupInfo 模型
 type GroupInfo struct {
 	ID        string `gorm:"column:id;primaryKey"` // 主键，字符串类型
-	CreatedAt int    `gorm:"column:created_at"`    // 创建时间
+	CreatedAt int64  `gorm:"column:created_at"`    // 创建时间
 	UpdatedAt *int64 `gorm:"column:updated_at"`    // 更新时间，int64类型
 	Data      []byte `gorm:"column:data"`          // BLOB 类型字段，使用 []byte 表示
 }
